fix setup.sh for pynq

+++ setup.sh
@@ -1,24 +1,25 @@
 #!/bin/bash
-sh -c '
-#On Raspbian Stretch we need to install old libssl (1.02 to be exact)
-sudo apt-get update
-sudo apt-get purge -y libssl-dev
-sudo apt-get install -y libssl1.0-dev git cmake build-essential curl libcurl4-openssl-dev uuid-dev
 
-#On Raspbian Stretch precompiled sdk from ppa is no good. Lets compile it manually
+if [ -z "$HOME" ]
+then
+      export HOME=/root
+fi
+curdir=$(pwd)
+
+
 cd ~
-mkdir Source
-cd Source
-git clone --recursive https://github.com/azure/azure-iot-sdk-c.git
+if [ ! -d ~/Source ]; then
+	mkdir Source
+	cd Source
+	git clone --recursive https://github.com/azure/azure-iot-sdk-c.git
 
-cd azure-iot-sdk-c/build_all/linux
-./build.sh --no-make
+	cd azure-iot-sdk-c/build_all/linux
+	./build.sh --no-make
+	cd ../../cmake/iotsdk_linux
+	make
+	sudo make install
+fi
 
-cd ../../cmake/iotsdk_linux
-make
-
-sudo make install
-'
 
 vercomp() {
     if [[ $1 == $2 ]]
@@ -71,20 +72,13 @@ else
     echo "cmake version check pass (current:$CMAKE_VER,require:$CMAKE_LEAST)"
 fi
 
-git clone git://git.drogon.net/wiringPi
-cd ./wiringPi
-./build
-cd ..
-
-git clone https://github.com/kgabis/parson.git
-cd ./parson
-mv parson.c parson.h ..
-cd ..
-rm -rf parson
-
-if [[ $1 == "--simulated-data" ]]; then
-    echo -e "Using simulated data"
-    sed -i 's/#define SIMULATED_DATA 0/#define SIMULATED_DATA 1/' config.h
+cd ${curdir}
+if [ ! -f parson.c ]; then
+	git clone https://github.com/kgabis/parson.git
+	cd ./parson
+	mv parson.c parson.h ..
+	cd ..
+	rm -rf parson
 fi
 
 cmake . && make
